% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{construct_ribotrans}
\alias{construct_ribotrans}
\title{Construct a ribotrans Object}
\usage{
construct_ribotrans(
  gtf_file = NULL,
  RNA_bam_file = NULL,
  RNA_sample_name = NULL,
  Ribo_bam_file = NULL,
  Ribo_sample_name = NULL,
  choose_longest_trans = FALSE
)
}
\arguments{
\item{gtf_file}{A \code{character} string indicating the path to the GTF annotation file.}

\item{RNA_bam_file}{A \code{character} vector containing paths to RNA-seq BAM files.}

\item{RNA_sample_name}{A \code{character} vector representing RNA-seq sample names.}

\item{Ribo_bam_file}{A \code{character} vector containing paths to ribosome profiling BAM files.}

\item{Ribo_sample_name}{A \code{character} vector representing ribosome profiling sample names.}

\item{choose_longest_trans}{A logical value indicating whether to select only
the longest transcript for each gene. Default is \code{FALSE}.}
}
\value{
An object of class \code{ribotrans} containing BAM file metadata,
library information, and parsed transcriptomic features.
}
\description{
Creates a \code{ribotrans} object from RNA and ribosome profiling BAM files along
with transcriptome annotation (\code{GTF} file).
}
\details{
This function processes the transcriptome annotation and extracts alignment
information from RNA and ribosome profiling BAM files to create
an S4 object of class \code{ribotrans}.

It computes total mapped reads for each BAM file and stores the information
in the \code{library} slot of the \code{ribotrans} object.
}
