% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_offset_correction}
\alias{do_offset_correction}
\title{Perform Offset Correction for Ribosome Footprints}
\usage{
do_offset_correction(object = NULL, shift = 0)
}
\arguments{
\item{object}{A \code{ribotrans} object.}

\item{shift}{Integer defining the \strong{offset shift} applied to read positions. \strong{Default}: \code{0}.}
}
\value{
A modified \code{data.frame} with updated \strong{ribosome footprint positions (\code{pos})}.
}
\description{
This function corrects \strong{ribosome footprint positions} based on \code{reads_offset_info},
adjusting \strong{relative positions} of mapped reads.
}
\details{
\itemize{
\item If \strong{reads_offset_info is missing}, an error message is displayed.
\item The offset correction is performed by adjusting \code{pos = pos - rel_pos + shift}.
}
}
\examples{
\dontrun{
# Apply offset correction
corrected_data <- do_offset_correction(obj, shift = 3)
head(corrected_data)
}


}
