% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_kmers}
\alias{generate_kmers}
\title{Generate Unique k-mer Peptides from Coding or Peptide Sequences}
\usage{
generate_kmers(cds_fa = NULL, peptide_fa = NULL, kmer_length = 15)
}
\arguments{
\item{cds_fa}{Character. File path to a FASTA file containing CDS (DNA sequences). If this is
provided, the sequences will be translated to peptides using \code{Biostrings::translate()}.}

\item{peptide_fa}{Character. File path to a FASTA file containing amino acid sequences. Used
only if \code{cds_fa} is NULL.}

\item{kmer_length}{Integer. Length of the peptide k-mers to be extracted. Default is 15.}
}
\value{
A character vector containing all unique amino acid k-mer subsequences
(length = \code{kmer_length}) from the translated or provided protein sequences.
}
\description{
This function extracts all unique amino acid k-mers (subsequences of fixed length) from
either DNA coding sequences (CDS) or peptide sequences. It supports automatic translation
from CDS to protein sequences using the standard genetic code and filters out invalid input
such as sequences with ambiguous nucleotides.
}
\details{
This function performs the following operations:
\itemize{
\item Reads CDS (DNA) or peptide sequences from a FASTA file.
\item Filters out DNA sequences containing ambiguous nucleotides (non-ACGT).
\item Translates CDS into amino acid sequences using the standard genetic code.
\item Selects sequences with length >= \code{kmer_length}.
\item Extracts overlapping k-mers from each sequence using a sliding window (step = 1).
\item Removes duplicate k-mers and returns a unique list.
}
Requires the \pkg{Biostrings} and \pkg{stringr} packages.
}
\examples{
\dontrun{
# Example: Generate 15-mers from CDS
kmers <- generate_kmers(cds_fa = "transcripts.fasta", kmer_length = 15)
head(kmers)

# Example: Generate 10-mers from amino acid sequences
kmers <- generate_kmers(peptide_fa = "proteins.fasta", kmer_length = 10)
head(kmers)
}

}
