% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aa_table}
\alias{get_aa_table}
\title{Generate a Codon to Amino Acid Translation Table}
\usage{
get_aa_table()
}
\value{
A data frame containing:
\item{Codon}{The three-letter codon sequence.}
\item{AminoAcid}{The full name of the corresponding amino acid.}
\item{Abbreviation3}{The three-letter abbreviation of the amino acid.}
\item{Abbreviation1}{The one-letter abbreviation of the amino acid.}
}
\description{
This function creates a codon-to-amino acid translation table based on the standard genetic code.
}
