% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{get_scaled_occupancy}
\alias{get_scaled_occupancy}
\title{Generic Function for Computing Scaled Ribosome Occupancy}
\usage{
get_scaled_occupancy(object, ...)
}
\arguments{
\item{object}{An object containing ribosome profiling (\code{ribo-seq}) and RNA-seq coverage data.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
The return structure depends on the method. Generally, it returns a \code{ribotrans} object with updated scaled occupancy data.
}
\description{
A S4 generic function to compute \strong{scaled ribosome occupancy}
(ribo-seq signal normalized to RNA-seq coverage).
This function dispatches S4 methods based on the class of \code{object}.
}
\details{
This generic function allows different implementations of \strong{scaled ribosome occupancy calculations}.
Specific implementations should be defined using \code{setMethod()}.
\itemize{
\item \strong{\verb{get_scaled_occupancy,ribotrans-method}}: Computes \strong{Ribo/RNA enrichment per position} from a \code{ribotrans} object.
}
}
\seealso{
\code{\link{get_occupancy}}, \code{\link{get_coverage}}
}
