% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-QC.R
\name{relative_heatmap_plot}
\alias{relative_heatmap_plot}
\alias{relative_heatmap_plot,ribotrans-method}
\title{Relative Heatmap Plot Generic Function}
\usage{
relative_heatmap_plot(object, ...)

\S4method{relative_heatmap_plot}{ribotrans}(
  object,
  type = c("rel2start", "rel2stop"),
  read_length = c(20, 35),
  rel_dist = c(-100, 100),
  log_scale = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{object}{An object of class \code{ribotrans}, containing summary information from ribosome profiling data.}

\item{...}{Additional arguments passed to specific methods.}

\item{type}{A character string indicating the type of plot. Options are:
\code{"rel2start"} to plot relative to start codon or
\code{"rel2stop"} to plot relative to stop codon. Default is \code{"rel2start"}.}

\item{read_length}{A numeric vector of length 2 specifying the range of read lengths to include in the plot. Default is \code{c(20, 35)}.}

\item{rel_dist}{A numeric vector of length 2 specifying the range of relative distances to plot. Default is \code{c(-100, 100)}.}

\item{log_scale}{A logical value indicating whether to apply a logarithmic transformation to counts. Default is \code{FALSE}.}

\item{return_data}{A logical value determining the return type. If \code{TRUE}, the summarized data is returned; if \code{FALSE}, the plot object is returned. Default is \code{FALSE}.}
}
\value{
A heatmap plot specific to the method invoked.

A ggplot object if \code{return_data} is \code{FALSE}, otherwise a data.table with summarized counts and relevant metadata.
}
\description{
This is a generic function for creating relative heatmap plots
across different data objects.

This function creates a heatmap that represents the relative position
of reads with respect to either start or stop codons of genes.
}
\details{
The function checks whether the summary information is available in the provided
\code{ribotrans} object and filters data based on the specified criteria.
The resulting heatmap visualizes the distribution of read counts along with
their relative positions to the coding features.
}
