% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{ribotrans-class}
\alias{ribotrans-class}
\alias{ribotrans}
\title{ribotrans Class}
\value{
An object of class \code{ribotrans}.
}
\description{
An S4 class to represent ribosome profiling data and associated information.
}
\details{
The \code{ribotrans} class is designed to facilitate the analysis of ribosome profiling data
by organizing relevant information, including BAM file metadata, gene annotations (GTF),
and various summary statistics related to ribosome and RNA-seq reads.

Key functionalities include:
\itemize{
\item Storing and accessing essential metadata
\item Aggregating ribosome profiling signals
\item Computing P-site offsets
\item Normalizing ribosome occupancy with RNA-seq coverage
}
}
\section{Slots}{

\describe{
\item{\code{bam_file}}{A \code{data.frame} containing BAM file data.}

\item{\code{library}}{A \code{data.frame} representing the sequencing library information.}

\item{\code{gtf_data}}{A \code{GRanges} representing the gtf data for gene annotations.}

\item{\code{gtf_path}}{A \code{character} string specifying the gtf file path.}

\item{\code{assignment_mode}}{A \code{character} string indicating the methodology used for read assignment.}

\item{\code{mapping_type}}{A \code{character} string specifying the mapping type (\code{"transcriptome"} or \code{"genome"}).}

\item{\code{features}}{A \code{data.frame} describing transcript features (e.g., exons, genes).}

\item{\code{genome_trans_features}}{A \code{data.frame} containing genomic transcript features for genome-based mappings.}

\item{\code{summary_info}}{A \code{data.frame} describing all reads detailed information.}

\item{\code{reads_offset_info}}{A \code{data.frame} recording information on ribosome P-site offsets.}

\item{\code{gene_name}}{A \code{character} string specifying the target gene name.}

\item{\code{ribo_occupancy}}{A \code{data.frame} storing ribosome occupancy data.}

\item{\code{ribo.smoothed}}{A \code{character} representing smoothed ribosome profiles.}

\item{\code{RNA_coverage}}{A \code{data.frame} containing RNA coverage data from sequencing.}

\item{\code{RNA.smoothed}}{A \code{character} representing smoothed RNA profiles.}

\item{\code{scaled_occupancy}}{A \code{data.frame} with normalized ribosome occupancy data.}

\item{\code{counts}}{A list for saving RPF and mRNA counts data.}
}}

