% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{smoothEachPosition}
\alias{smoothEachPosition}
\title{Smooth Data for Each Transcript Position}
\usage{
smoothEachPosition(features = NULL, posdf = NULL, slide_window = 30)
}
\arguments{
\item{features}{A \code{data.frame} containing transcript annotations.}

\item{posdf}{A \code{data.frame} containing positional read count information with columns:
\describe{
\item{rname}{Character: Transcript name in the format \code{"transcript_id|gene_name"}.}
\item{pos}{Integer: Position along the transcript where reads are mapped.}
\item{sample}{Character: Sample identifier.}
\item{rpm}{Numeric: Reads per million mapped reads at each position.}
}}

\item{slide_window}{An \code{integer} defining the window size for the rolling mean (default: \code{30}).}
}
\value{
A \code{data.frame} containing smoothed transcript positions with columns:
\describe{
\item{rname}{Character: Transcript name (\code{transcript_id|gene_name}).}
\item{pos}{Integer: Position within the transcript.}
\item{sample}{Character: Sample name.}
\item{rpm}{Numeric: Original reads per million (RPM) values.}
\item{smooth}{Numeric: Smoothed RPM values using a rolling mean filter.}
}
}
\description{
Applies a sliding window smoothing function to mapped read counts (RPM) across transcript positions.
}
\details{
This function extracts transcript-specific mapped read counts from \code{posdf},
aligns them with transcript annotations in \code{features}, and applies a moving
average smoothing function (\code{zoo::rollmean}) using a \strong{sliding window approach}.

It processes each transcript separately, ensuring that all positions (1 to
exon length) are covered, filling missing values with \code{0} before smoothing.
}
