% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-data.R
\name{subset_data}
\alias{subset_data}
\alias{subset_data,ribotrans-method}
\alias{subset_data,serp-method}
\title{Subset Data in \code{serp} and \code{ribotrans} Objects}
\usage{
subset_data(object, ...)

\S4method{subset_data}{ribotrans}(object, ...)

\S4method{subset_data}{serp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{serp} or \code{ribotrans}.}

\item{...}{Logical conditions used for subsetting data frames inside the object.
These are passed to the standard \code{subset()} function.}
}
\value{
An updated object of class \code{serp} or \code{ribotrans}, with internal slots filtered
according to the given criteria.
}
\description{
A generic method to subset \code{serp} and \code{ribotrans} objects. This function filters internal data
slots (\code{bam_file}, \code{library}, \code{summary_info}) based on user-defined criteria.
}
\details{
For \code{ribotrans} objects, this method:
\itemize{
\item Subsets the \code{bam_file} data frame based on \code{...}.
\item Subsets the \code{library} data frame based on \code{...}.
\item Filters \code{summary_info} to include only rows where \code{sample} exists in the filtered \code{library}.
}

For \code{serp} objects, this method:
\itemize{
\item Subsets the \code{bam_file} data frame based on \code{...}.
\item Subsets the \code{library} data frame based on \code{...}.
\item Filters \code{summary_info} where the \code{sample} column matches
\code{paste(lib$sample, lib$type, sep = "-")}, typically used for compound sample IDs.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{subset_data(ribotrans)}: Method for ribotrans objects

\item \code{subset_data(serp)}: Method for serp objects

}}
\examples{
\dontrun{
# Subset a ribotrans object to include only a specific sample
subset_data(ribo, sample == "SampleA")

# Subset a serp object based on type in library
subset_data(serp_obj, type == "ip")
}

}
