% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-visualizations.R
\name{vocalno_plot}
\alias{vocalno_plot}
\title{Volcano Plot for Differential Gene Expression Results}
\usage{
vocalno_plot(
  diff_data = NULL,
  log2FC = "log2FoldChange",
  pval = "pvalue",
  log2FC_threshold = c(-1, 1),
  top_gene = 10,
  marker_gene = NULL,
  gene_number_size = 4,
  gene_number_label_pos = c(0.95, 0.95),
  gene_label_size = 3,
  color = c("#AF1740", "grey", "#074799")
)
}
\arguments{
\item{diff_data}{A data frame containing the differential expression results. Must include columns for log2 fold change, p-value, and a categorical 'type' column (e.g., "sigUp", "sigDown", "nonSig") indicating significance status. A 'gene_name' column is also required for labeling.}

\item{log2FC}{String. The column name representing log2 fold changes. Default is \code{"log2FoldChange"}.}

\item{pval}{String. The column name representing the p-value. Default is \code{"pvalue"}.}

\item{log2FC_threshold}{Numeric vector of length 2. Thresholds for determining significant up/down regulation based on log2 fold change. Default is \code{c(-1, 1)}.}

\item{top_gene}{Integer. Number of most significant upregulated/downregulated genes (based on p-value) to label if \code{marker_gene} is not provided. Default is \code{10}.}

\item{marker_gene}{Character vector. Optional list of gene names to highlight on the plot. If NULL, top genes are selected based on significance and p-value. Default is \code{NULL}.}

\item{gene_number_size}{Numeric. Text size for displaying significant gene counts. Default is \code{4}.}

\item{gene_number_label_pos}{Numeric vector of length 2. Position (npc coordinates) to place gene count labels. Default is \code{c(0.95, 0.95)}.}

\item{gene_label_size}{Numeric. Text size for gene name labels. Default is \code{3}.}

\item{color}{Character vector of length 3. Colors used for sigUp, nonSig, and sigDown categories. Default is \code{c("#AF1740", "grey", "#074799")}.}
}
\value{
A \code{ggplot} object representing the volcano plot.
}
\description{
Generate a volcano plot to visualize differentially expressed genes based on log2 fold change and p-value. Highlight significantly upregulated and downregulated genes, optionally label top genes or specific marker genes, and display counts of significant genes on the plot.
}
\details{
The function draws a volcano plot using ggplot2. Points are colored by significance category (\code{type}), with cutoff thresholds visualized as dashed lines. It also annotates the number of significant genes and optionally labels key genes. Top significant genes are determined based on lowest p-value unless \code{marker_gene} is specified.
}
\examples{
\dontrun{
vocalno_plot(diff_data = my_diff_data,
             top_gene = 10,
             marker_gene = c("RPL3", "ACT1"),
             gene_label_size = 3.5)
}

}
