% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{generate_summary}
\alias{generate_summary}
\title{Generate Summary Information for Ribosome Profiling Data}
\usage{
generate_summary(object, ...)
}
\arguments{
\item{object}{An object containing experimental data}

\item{...}{Additional arguments passed to methods}
}
\value{
An updated object containing summary information
}
\description{
Generic function to generate summary information from ribosome profiling
or RNA-seq data.
}
