% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation-genome.R
\name{genome_trans_plot}
\alias{genome_trans_plot}
\alias{genome_trans_plot,ribotrans-method}
\title{Generic function for genome transcript plotting}
\usage{
genome_trans_plot(object, ...)

\S4method{genome_trans_plot}{ribotrans}(
  object,
  selected_id = NULL,
  type = c("ribo", "rna", "ribo_rna"),
  layer = c("col", "line"),
  sample_order = NULL,
  facet_layer = ggplot2::facet_grid(sample ~ rname, switch = "y"),
  sep_mer_sample = FALSE,
  new_signal_range = FALSE,
  collapse_structure = FALSE,
  range_x = 0.9,
  range_y = 0.9,
  range_size = 4,
  range_digit = 1,
  scale_factor = 1,
  utr_width = 1,
  cds_width = 3,
  background_line = c("black", 0.25),
  exon_line = c("#003399", 1),
  cds_line = c("#003399", 3),
  structure_panel_height = 0.2
)
}
\arguments{
\item{object}{An object of class \code{ribotrans} containing the necessary data for plotting.}

\item{...}{Additional arguments passed to specific methods.}

\item{selected_id}{(optional) A specific transcript id for the gene of interest. Default is NULL.}

\item{type}{A character string indicating the type of data to plot. Choices are "ribo" for
ribosomal occupancy, "rna" for RNA coverage, or "ribo_rna" for both. Default is "ribo".}

\item{layer}{A character string indicating the plot layer type, either "col" for \code{geom_col} plots
or "line" for \code{geom_path} plots. Default is "col".}

\item{sample_order}{(optional) A character vector specifying the order in which samples should be
displayed in the plot. Default is NULL.}

\item{facet_layer}{A facet layer for ggplot, default is \code{ggplot2::facet_grid(sample~rname,switch = "y")}.}

\item{sep_mer_sample}{A boolean indicating whether to separate merged and individual samples. Default is FALSE.}

\item{new_signal_range}{A boolean indicating whether to show the new signal range on the plot. Default is FALSE.}

\item{collapse_structure}{A boolean indicating whether to collapse the structures of transcripts and exons into a single line in the plot. Default is FALSE.}

\item{range_x}{A numeric value specifying the position adjustment for the range label on the x-axis. Default is 0.9.}

\item{range_y}{A numeric value specifying the position adjustment for the range label on the y-axis. Default is 0.9.}

\item{range_size}{A numeric value specifying the font size for the range label. Default is 4.}

\item{range_digit}{An integer indicating the number of digits to round the range label. Default is 1.}

\item{scale_factor}{A scaling factor to adjust RNA coverage values for better visualization. Default is 1.}

\item{utr_width}{A numeric value to specify the width of the UTR segments in the plot. Default is 1.}

\item{cds_width}{A numeric value to specify the width of the CDS segments in the plot. Default is 3.}

\item{background_line}{A vector defining the color and linewidth for the background line structure. Default is \code{c("black", 0.25)}.}

\item{exon_line}{A vector defining the color and linewidth for the exon line structure. Default is \code{c("#003399", 1)}.}

\item{cds_line}{A vector defining the color and linewidth for the CDS line structure. Default is \code{c("#003399", 3)}.}

\item{structure_panel_height}{A numeric value specifying the height of the structure panel in the plot. Default is 0.2.}
}
\value{
A ggplot object or other relevant output depending on the method used.

A ggplot object displaying the genomic transcripts along with ribosome occupancy or RNA coverage data.
}
\description{
This generic function serves as an interface for plotting genomic transcripts.
It allows specific methods to be applied depending on the class of the object provided.

This method generates genomic plots for ribosomal occupancy and RNA reads coverage
based on the specified parameters for objects of class \code{ribotrans}.
}
\details{
The function uses ggplot2 for visualization and requires specific data to be pre-calculated (e.g., ribosome occupancy or RNA coverage). Users must ensure that the appropriate methods (such as \code{get_occupancy} or \code{get_coverage}) have been executed prior to plotting.
}
\examples{
\dontrun{
# Assuming 'ribotrans_obj' is a pre-defined ribotrans object
genome_trans_plot(
  object = ribotrans_obj,
  type = "ribo",
  layer = "col",
  sample_order = c("sample1", "sample2", "merged_sample"),
  collapse_structure = TRUE
)
}

}
