% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{get_occupancy}
\alias{get_occupancy}
\title{Generic Function for Extracting Ribosome Occupancy}
\usage{
get_occupancy(object, ...)
}
\arguments{
\item{object}{An object containing ribosome profiling (\code{ribo-seq}) or transcriptomic data.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
The output format depends on the method. Generally, it returns a \code{ribotrans} object with updated occupancy data.
}
\description{
A S4 generic function for extracting ribosome footprinting occupancy from different object types.
It dispatches S4 methods based on the class of \code{object}.
}
\details{
This generic function enables \strong{ribosome occupancy extraction} from different S4 objects.
Specific implementations are handled via \code{setMethod()}.
\itemize{
\item \strong{\verb{get_occupancy,ribotrans-method}}: Extracts ribosome footprinting per-position occupancy from a \code{ribotrans} object.
}
}
\seealso{
\code{\link{getOccupancy}} for computing occupancy directly from BAM files.
}
