% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-methods.R
\name{length_plot}
\alias{length_plot}
\alias{length_plot,ribotrans-method}
\title{Generate Length Distribution Plot}
\usage{
length_plot(object, ...)

\S4method{length_plot}{ribotrans}(object, return_data = FALSE, type = c("length", "frame_length"))
}
\arguments{
\item{object}{A \code{ribotrans} object containing summary information.}

\item{...}{Additional arguments passed to class-specific methods.}

\item{return_data}{Whether return the data for plot, default \code{FALSE}.}

\item{type}{Character string specifying the plot type.
It must be one of the following:
\itemize{
\item \code{"length"} - Produces a histogram of read lengths.
\item \code{"frame_length"} - Produces a histogram of read lengths, colored by reading frame.
}}
}
\value{
A ggplot object representing the length distribution of reads.

A \code{ggplot2} object visualizing either the total read length distribution or frame-specific
read lengths.
}
\description{
This generic function generates a length distribution plot for sequencing reads.

This method creates a plot showing read length distributions or read-frame distributions
for sequencing data stored in a \code{ribotrans} object.
}
\details{
\itemize{
\item \code{"length"}: Groups reads by length (\code{qwidth}) and plots their total counts.
\item \code{"frame_length"}: Groups reads by frame (\code{(pos - mstart) \%\% 3}) and colors them accordingly.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{length_plot(ribotrans)}: Generate a length distribution or frame-length plot for \code{ribotrans} objects.

}}
