% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-QC.R
\name{metagene_plot}
\alias{metagene_plot}
\alias{metagene_plot,ribotrans-method}
\title{Generate Metagene Profile Plot}
\usage{
metagene_plot(object, ...)

\S4method{metagene_plot}{ribotrans}(
  object,
  type = c("rel2start", "rel2stop"),
  return_data = FALSE,
  mode = c("nt", "codon"),
  read_length = c(25, 31),
  rel2st_dist = c(-50, 100),
  rel2sp_dist = c(-100, 50),
  facet_wrap = ggplot2::facet_wrap(~sample)
)
}
\arguments{
\item{object}{A \code{ribotrans} object containing RNA sequencing metrics.}

\item{...}{Additional arguments passed to specific methods.}

\item{type}{Character string specifying the metagene type. Must be one of:
\itemize{
\item \code{"rel2start"} - Distance relative to the start codon.
\item \code{"rel2stop"} - Distance relative to the stop codon.
}}

\item{return_data}{Logical, if \code{TRUE}, returns the processed data instead of a plot (default: \code{FALSE}).}

\item{mode}{Character string specifying the unit for the x-axis:
\itemize{
\item \code{"nt"}: Nucleotide level resolution (default).
\item \code{"codon"}: Codon-level resolution.
}}

\item{read_length}{Numeric vector of length two specifying the range of read lengths to include (default: \code{c(25,31)}).}

\item{rel2st_dist}{Numeric vector of length two, specifying distance range to include relative to start codon (default: \code{c(-50,100)}).}

\item{rel2sp_dist}{Numeric vector of length two, specifying distance range to include relative to stop codon (default: \code{c(-100,50)}).}

\item{facet_wrap}{A \code{ggplot2::facet_wrap()} object to control faceting by sample (default behavior).}
}
\value{
A \code{ggplot2} object showing the metagene profile distribution.

If \code{return_data = FALSE}, returns a \code{ggplot2} object showing the metagene profile plot.
If \code{return_data = TRUE}, returns a processed \code{data.frame} containing metagene read distributions.
}
\description{
A generic function to generate a metagene profile plot
visualizing read distributions relative to start or stop codons.

This method creates a metagene profile plot showing the
distribution of ribosome-protected fragments (RPFs) relative to the
start or stop codon.
}
\details{
\itemize{
\item The function normalizes read counts based on total mapped reads (RPM normalization).
\item Optionally, it applies a rolling average (\code{slide_window}) to smooth the profile.
\item If \code{show_frame = TRUE}, colors are applied for individual reading frames.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{metagene_plot(ribotrans)}: Generate a metagene profile plot for \code{ribotrans} objects.

}}
\examples{
\dontrun{
# Generate standard metagene plot relative to start codon
metagene_plot(ribo_obj, type = "rel2start")


# Get the underlying data instead of a plot
data <- metagene_plot(ribo_obj, type = "rel2start", return_data = TRUE)
head(data)
}

}
