% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-methods.R
\name{trans_plot}
\alias{trans_plot}
\title{Visualize transcript signals}
\usage{
trans_plot(object, ...)
}
\arguments{
\item{object}{The main object containing required data for visualization. Typically an S4 class object, such as \code{ribotrans}.}

\item{...}{Additional parameters to be used in specific method implementations.}
}
\description{
Generic function definition for visualizing transcript-level signals, including ribosome occupancy and RNA-seq coverage.
}
\seealso{
\code{\link{trans_plot,ribotrans-method}} for the method implementation specific to the \code{ribotrans} class.
}
