% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp-visualization.R
\name{trans_plot}
\alias{trans_plot}
\title{Plot translation profiles}
\usage{
trans_plot(object, ...)
}
\arguments{
\item{object}{An object of class \code{serp} or \code{ribotrans}, containing necessary data for plotting.}

\item{...}{Additional arguments passed to the specific method.}
}
\value{
A plot, typically a ggplot2 object.
}
\description{
This is a generic function for plotting translation-related profiles (such as
ribosome occupancy or enrichment signals) for different types of transcriptome data.
}
